+++ mintapi/constants.py
@@ -1,7 +1,11 @@
 JSON_FORMAT = "json"
 CSV_FORMAT = "csv"
 
+GET_METHOD = "GET"
+POST_METHOD = "POST"
+
 ACCOUNT_KEY = "Account"
+BILL_KEY = "Bill"
 BUDGET_KEY = "Budget"
 CATEGORY_KEY = "Category"
 INVESTMENT_KEY = "Investment"
@@ -9,6 +13,7 @@ TRANSACTION_KEY = "Transaction"
 NET_WORTH_KEY = "Net_Worth"
 CREDIT_SCORE_KEY = "Credit_Score"
 CREDIT_REPORT_KEY = "Credit_Report"
+TRENDS_KEY = "Trend"
 
 MINT_ROOT_URL = "https://mint.intuit.com"
 MINT_BETA_ROOT_URL = "https://beta.mint.intuit.com"